Separated error alert component for reusability
To be used in other widgets, based on the
